Some improvements to the table-aliases
- changed the various "con." alias to "cn." for consistency
- removed the undeclared "con." alias from the linkpoint file because it was causing an error
- In a few places removed the unnecessary "c." prefixes just to minimize code changes from original core code.
  (ie: the "c." or "cn." prefix only needs to be mentioned when there are other fields of the same name which could create ambiguity, especially in SELECT and ORDER BY clauses. Sometimes it's helpful to mention them anyway just for clarity to the coder, and that's fine.

I also notice that the "countries_name" field in the COUNTRIES table is left intact, which does create some ambiguity about which table contains the master copy of the names. If this were to be considered for core-code then that ambiguity would need to be mitigated, ideally by deleting the column from the COUNTRIES table.